Sending lux values over ESPNOW packets

+++ IMEG_ALL_Meter-main/Firmware/ALLRemote/remote_packet.h
@@ -1,4 +1,4 @@
 struct remote_packet {
-  int lux = 45;
+  uint16_t lux = 45;
   int lat = 90;
 };
